requirements/base.txt
pytz==2022.6  # https://github.com/stub42/pytz
redis==4.3.4  # https://github.com/redis/redis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.19.0  # https://github.com/encode/uvicorn
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
Pillow

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.24.2  # https://github.com/tfranzel/drf-spectacular

#custom
djangorestframework-role-filters==1.0.1
rest-framework-cache==0.1
fcm-django==1.0.12
django-filter==22.1
drf-api-logger==1.1.11
django-soft-delete==0.9.21
